requirements.txt
torch
torchaudio
coqui-tts
gradio
sounddevice
noisereduce
openai-whisper
